requirements.txt
Django==4.2.9
wagtail==5.1
wagtail-modeladmin==1.0.0
Pillow==10.1.0
python-slugify==8.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.12.0
django-extensions==3.2.3
requests==2.31.0
